Update build_database.ps1
Gone over the build script moving some of the variables that had been put into the code in early development to the top of the script.

Unified all of the error handling, anything that was doing something has now been put inside a try catch block with error catching to alert when something goes wrong.

The script will now stop and error if something fundamental to the build fails.

Each of the dataUpdates have been put into their own invoke try, this is because they need to occur in a particular order, this needs looking into to see if there is a cleaner way of doing this going forward.